Resolved issue: NETTY-298 Make ChannelBuffer setter/writer methods to accept int for narrower types

+++ src/main/java/org/jboss/netty/buffer/BigEndianHeapChannelBuffer.java
@@ -88,7 +88,7 @@ public class BigEndianHeapChannelBuffer extends HeapChannelBuffer {
                 ((long) array[index+7] & 0xff) <<  0;
     }
 
-    public void setShort(int index, short value) {
+    public void setShort(int index, int value) {
         array[index  ] = (byte) (value >>> 8);
         array[index+1] = (byte) (value >>> 0);
     }
